+++ CHANGELOG.md
@@ -7,8 +7,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ### Added
 
-- Allow deriving child based on 32-byte digest.
-- Add fallible version of `SecretTree::fill` called `SecretTree::try_fill`.
+- Allow deriving a child from a tree based on 32-byte digest.
+- Add fallible versions of `SecretTree::fill()` and `SecretTree::create_secret()`
+  called `SecretTree::try_fill()` and `SecretTree::try_create_secret()`,
+  respectively.
 
 ### Changed