+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+## 0.5.0 - 2022-07-21
+
 ### Added
 
 - Allow deriving a child from a tree based on 32-byte digest.